herve/src/main/webapp/app/entities/sensor-type/route/sensor-type-routing-resolve.service.spec.ts
import { TestBed } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { ActivatedRouteSnapshot, ActivatedRoute, Router, convertToParamMap } from '@angular/router';
import { RouterTestingModule } from '@angular/router/testing';
import { of } from 'rxjs';

import { ISensorType, SensorType } from '../sensor-type.model';
import { SensorTypeService } from '../service/sensor-type.service';

import { SensorTypeRoutingResolveService } from './sensor-type-routing-resolve.service';

describe('SensorType routing resolve service', () => {
  let mockRouter: Router;
  let mockActivatedRouteSnapshot: ActivatedRouteSnapshot;
  let routingResolveService: SensorTypeRoutingResolveService;
  let service: SensorTypeService;
  let resultSensorType: ISensorType | undefined;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule, RouterTestingModule.withRoutes([])],
      providers: [
        {
          provide: ActivatedRoute,
          useValue: {
            snapshot: {
              paramMap: convertToParamMap({}),
            },
          },
        },
      ],
    });
    mockRouter = TestBed.inject(Router);
    jest.spyOn(mockRouter, 'navigate').mockImplementation(() => Promise.resolve(true));
    mockActivatedRouteSnapshot = TestBed.inject(ActivatedRoute).snapshot;
    routingResolveService = TestBed.inject(SensorTypeRoutingResolveService);
    service = TestBed.inject(SensorTypeService);
    resultSensorType = undefined;
  });

  describe('resolve', () => {
    it('should return ISensorType returned by find', () => {
      // GIVEN
      service.find = jest.fn(id => of(new HttpResponse({ body: { id } })));
      mockActivatedRouteSnapshot.params = { id: 123 };

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultSensorType = result;
      });

      // THEN
      expect(service.find).toBeCalledWith(123);
      expect(resultSensorType).toEqual({ id: 123 });
    });

    it('should return new ISensorType if id is not provided', () => {
      // GIVEN
      service.find = jest.fn();
      mockActivatedRouteSnapshot.params = {};

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultSensorType = result;
      });

      // THEN
      expect(service.find).not.toBeCalled();
      expect(resultSensorType).toEqual(new SensorType());
    });

    it('should route to 404 page if data not found in server', () => {
      // GIVEN
      jest.spyOn(service, 'find').mockReturnValue(of(new HttpResponse({ body: null as unknown as SensorType })));
      mockActivatedRouteSnapshot.params = { id: 123 };

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultSensorType = result;
      });

      // THEN
      expect(service.find).toBeCalledWith(123);
      expect(resultSensorType).toEqual(undefined);
      expect(mockRouter.navigate).toHaveBeenCalledWith(['404']);
    });
  });
});
